improve junit tests
